requirements.txt
amqp==5.0.2
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.5
celery-progress==0.0.14
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.5
django-celery-results==2.0.0
django-cors-headers==3.6.0
django-cors-middleware==1.5.0
djangorestframework==3.12.2
idna==2.10
importlib-metadata==3.4.0
kombu==5.0.2
prompt-toolkit==3.0.10
pytz==2020.5
redis==3.2.1
requests==2.25.1
selenium==3.141.0
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.0
